custom-nodes/node-red-contrib-crowdcountingnode/crowdCountingNode/resources/requirements.txt
opencv-python
Pillow
numpy
torch
torchvision